+++ core/apis.go
@@ -145,16 +145,20 @@ func (s *administrationImpl) AdmGetApplications() ([]model.Application, error) {
 	return s.app.admGetApplications()
 }
 
-func (s *administrationImpl) AdmCreateApplicationPermission(name string, appID string) (*model.ApplicationPermission, error) {
-	return s.app.admCreateApplicationPermission(name, appID)
+func (s *administrationImpl) AdmCreatePermission(name string, serviceIDs []string) (*model.Permission, error) {
+	return s.app.admCreatePermission(name, serviceIDs)
+}
+
+func (s *administrationImpl) AdmUpdatePermission(name string, serviceIDs *[]string) (*model.Permission, error) {
+	return s.app.admUpdatePermission(name, serviceIDs)
 }
 
 func (s *administrationImpl) AdmCreateApplicationRole(name string, appID string, description string, permissionNames []string) (*model.ApplicationRole, error) {
 	return s.app.admCreateApplicationRole(name, appID, description, permissionNames)
 }
 
-func (s *administrationImpl) AdmGrantAccountPermissions(accountID string, appID string, permissionNames []string) error {
-	return s.app.admGrantAccountPermissions(accountID, appID, permissionNames)
+func (s *administrationImpl) AdmGrantAccountPermissions(accountID string, permissionNames []string) error {
+	return s.app.admGrantAccountPermissions(accountID, permissionNames)
 }
 
 func (s *administrationImpl) AdmGrantAccountRoles(accountID string, appID string, roleIDs []string) error {